+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Add or remove tags from Containers list and details page
 - Show tags on containers
 - Add "Cannery" to page titles
+- Don't show true/false column for ammo types if all values are false
 
 
 # 0.1.0